--- src/GoalList.css
@@ -1,15 +1,0 @@
-.GoalList {
-  height: 10em;
-  overflow-y: scroll;
-
-  font-size: medium;
-  
-}
-
-.GoalList .close-button {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  
-  padding-right: .5em;
-}
